Export menu to a separate file
