+++ BookUniverse.Client/CustomControls/Menu.xaml.cs
@@ -1,0 +1,27 @@
+﻿using System.Windows;
+using System.Windows.Controls;
+namespace BookUniverse.Client.CustomControls
+{
+    /// <summary>
+    /// Interaction logic for Menu.xaml
+    /// </summary>
+    public partial class Menu : UserControl
+    {
+        public Menu()
+        {
+            InitializeComponent();
+        }
+
+        private void ButtonOpenMenu_Click(object sender, RoutedEventArgs e)
+        {
+            ButtonCloseMenu.Visibility = Visibility.Visible;
+            ButtonOpenMenu.Visibility = Visibility.Collapsed;
+        }
+
+        private void ButtonCloseMenu_Click(object sender, RoutedEventArgs e)
+        {
+            ButtonCloseMenu.Visibility = Visibility.Collapsed;
+            ButtonOpenMenu.Visibility = Visibility.Visible;
+        }
+    }
+}
