+++ BookUniverse.Client/Views/UserAccount.xaml
@@ -4,29 +4,8 @@
         xmlns:d="http://schemas.microsoft.com/expression/blend/2008"
         xmlns:mc="http://schemas.openxmlformats.org/markup-compatibility/2006"
         xmlns:materialDesign="http://materialdesigninxaml.net/winfx/xaml/themes"
+        xmlns:customcontrols="clr-namespace:BookUniverse.Client.CustomControls"
         mc:Ignorable="d" Height="750" Width="1400" Foreground="#FFE3E0" ResizeMode="NoResize" WindowStartupLocation="CenterScreen" WindowStyle="None" AllowsTransparency="True" Background="Transparent">
-    <Window.Resources>
-        <Storyboard x:Key="OpenMenu">
-            <DoubleAnimationUsingKeyFrames Storyboard.TargetProperty="(FrameworkElement.Width)" Storyboard.TargetName="GridMenu">
-                <EasingDoubleKeyFrame KeyTime="0" Value="70"/>
-                <EasingDoubleKeyFrame KeyTime="0:0:0.5" Value="200"/>
-            </DoubleAnimationUsingKeyFrames>
-        </Storyboard>
-        <Storyboard x:Key="CloseMenu">
-            <DoubleAnimationUsingKeyFrames Storyboard.TargetProperty="(FrameworkElement.Width)" Storyboard.TargetName="GridMenu">
-                <EasingDoubleKeyFrame KeyTime="0" Value="200"/>
-                <EasingDoubleKeyFrame KeyTime="0:0:0.5" Value="70"/>
-            </DoubleAnimationUsingKeyFrames>
-        </Storyboard>
-    </Window.Resources>
-    <Window.Triggers>
-        <EventTrigger RoutedEvent="ButtonBase.Click" SourceName="ButtonOpenMenu">
-            <BeginStoryboard Storyboard="{StaticResource OpenMenu}"/>
-        </EventTrigger>
-        <EventTrigger RoutedEvent="ButtonBase.Click" SourceName="ButtonCloseMenu">
-            <BeginStoryboard Storyboard="{StaticResource CloseMenu}"/>
-        </EventTrigger>
-    </Window.Triggers>
     <Grid Background="#FFE3E0">
         <Grid.ColumnDefinitions>
             <ColumnDefinition Width="300*"/>
@@ -46,13 +25,9 @@
             </StackPanel>
         </Grid>
 
-        <!-- Right Side -->
         <Grid Width="400" Margin="0 0" HorizontalAlignment="Right">
-            <!-- Add your picture here -->
             <Image Source="\Images\book_universe2.jpg" Margin="102,145,-325,73" />
         </Grid>
-
-        <Grid x:Name="GridMain" HorizontalAlignment="Right" VerticalAlignment="Bottom" Height="540" Width="1010" Grid.ColumnSpan="2"/>
         <Grid Height="60" VerticalAlignment="Top" Background="#2E294E" Grid.ColumnSpan="2">
             <Grid.Effect>
                 <DropShadowEffect BlurRadius="10" ShadowDepth="0" Color="#FFE3E0"/>
@@ -71,77 +46,6 @@
             </StackPanel>
             <TextBlock HorizontalAlignment="Center" VerticalAlignment="Center" FontSize="22">Book Universe</TextBlock>
         </Grid>
-        <Grid x:Name="GridMenu" Width="70" HorizontalAlignment="Left" Background="#2E294E">
-            <Grid.Effect>
-                <DropShadowEffect BlurRadius="10" ShadowDepth="0" Color="#011638"/>
-            </Grid.Effect>
-            <StackPanel>
-                <Border CornerRadius="10">
-                    <Grid Height="60" Background="White" RenderTransformOrigin="0.5,0.5">
-                        <Grid.Effect>
-                            <DropShadowEffect BlurRadius="10" ShadowDepth="0" Color="#011638"/>
-                        </Grid.Effect>
-                        <Grid.RenderTransform>
-                            <TransformGroup>
-                                <ScaleTransform/>
-                                <SkewTransform/>
-                                <RotateTransform/>
-                                <TranslateTransform/>
-                            </TransformGroup>
-                        </Grid.RenderTransform>
-                        <Button x:Name="ButtonOpenMenu" Height="60" Width="60" Margin="5" HorizontalAlignment="Right" VerticalAlignment="Top" Style="{StaticResource MaterialDesignFloatingActionMiniAccentButton}" Background="{x:Null}" BorderBrush="{x:Null}" Foreground="#FF31577E" Click="ButtonOpenMenu_Click">
-                            <materialDesign:PackIcon Kind="Menu" Width="25" Height="25"/>
-                        </Button>
-                        <Button x:Name="ButtonCloseMenu" Visibility="Collapsed" Height="60" Width="60" Margin="5" HorizontalAlignment="Right" VerticalAlignment="Top" Style="{StaticResource MaterialDesignFloatingActionMiniAccentButton}" Background="{x:Null}" BorderBrush="{x:Null}" Foreground="#FF31577E" Click="ButtonCloseMenu_Click">
-                            <materialDesign:PackIcon Kind="ArrowLeft" Width="25" Height="25"/>
-                        </Button>
-
-                        <Image Height="10" VerticalAlignment="Bottom"/>
-                    </Grid>
-                </Border>
-                <ListView x:Name="ListViewMenu" Foreground="#FF5C99D6" ScrollViewer.HorizontalScrollBarVisibility="Disabled">
-                    <ListViewItem x:Name="ItemHome" Height="60">
-                        <StackPanel Orientation="Horizontal">
-                            <materialDesign:PackIcon Kind="Bookshelf" Height="25" Width="25" Margin="10"/>
-                            <TextBlock Text="All books" VerticalAlignment="Center" Margin="20 10"/>
-                        </StackPanel>
-                    </ListViewItem>
-                    <ListViewItem x:Name="ItemCreate" Height="60">
-                        <StackPanel Orientation="Horizontal">
-                            <materialDesign:PackIcon Kind="Category" Height="25" Width="25" Margin="10"/>
-                            <ComboBox Width="115" Margin="20 10 0 10" materialDesign:HintAssist.HelperText="Categories" materialDesign:TextFieldAssist.UnderlineBrush="#182043" >
-                                <ComboBoxItem IsSelected="True" Content="Option 1" />
-                                <ComboBoxItem Content="Option 2" />
-                                <ComboBoxItem Content="Option 3" />
-                            </ComboBox>
-                        </StackPanel>
-                    </ListViewItem>
-
-
-                    <ListViewItem Height="60">
-                        <StackPanel Orientation="Horizontal">
-                            <materialDesign:PackIcon Kind="Favorite" Height="25" Width="25" Margin="10"/>
-                            <TextBlock Text="Favourites" VerticalAlignment="Center"  Margin="20 10"/>
-                        </StackPanel>
-                    </ListViewItem>
-                    <ListViewItem Height="60">
-                        <StackPanel Orientation="Horizontal">
-                            <materialDesign:PackIcon Kind="Folder" Height="25" Width="25" Margin="10"/>
-                            <ComboBox Width="115" Margin="20 10 0 10" materialDesign:HintAssist.HelperText="Folders" materialDesign:TextFieldAssist.UnderlineBrush="#182043" >
-                                <ComboBoxItem IsSelected="True" Content="Option 1" />
-                                <ComboBoxItem Content="Option 2" />
-                                <ComboBoxItem Content="Option 3" />
-                            </ComboBox>
-                        </StackPanel>
-                    </ListViewItem>
-                    <ListViewItem Height="60">
-                        <StackPanel Orientation="Horizontal">
-                            <materialDesign:PackIcon Kind="Search" Height="25" Width="25" Margin="10"/>
-                            <TextBlock Text="Search" VerticalAlignment="Center"  Margin="20 10"/>
-                        </StackPanel>
-                    </ListViewItem>
-                </ListView>
-            </StackPanel>
-        </Grid>
+        <customcontrols:Menu/>
     </Grid>
 </Window>
