+++ BookUniverse.Client/Views/UserAccount.xaml.cs
@@ -28,18 +28,6 @@ namespace BookUniverse.Client
             InitializeComponent();
         }
 
-        private void ButtonOpenMenu_Click(object sender, RoutedEventArgs e)
-        {
-            ButtonCloseMenu.Visibility = Visibility.Visible;
-            ButtonOpenMenu.Visibility = Visibility.Collapsed;
-        }
-
-        private void ButtonCloseMenu_Click(object sender, RoutedEventArgs e)
-        {
-            ButtonCloseMenu.Visibility = Visibility.Collapsed;
-            ButtonOpenMenu.Visibility = Visibility.Visible;
-        }
-
         private void CloseWindow(object sender, RoutedEventArgs e)
         {
             Application.Current.MainWindow.Close();
